pyproject.toml
[project]
name = "hid2uart"
version = "0.1.0"
description = "HID to UART bridge - capture keyboard, mouse, and gamepad input on Windows and transmit over serial port"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "HID2UART Project" },
]
keywords = ["hid", "uart", "serial", "keyboard", "mouse", "gamepad", "bridge", "pyqt5"]

dependencies = [
    "PyQt5>=5.15,<6.0",
    "pyserial>=3.5,<4.0",
    "pynput>=1.7,<2.0",
    "pygame>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "ruff>=0.1.0,<0.9.0",
    "mypy>=1.0,<1.15",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ui*", "input*", "utils*"]
namespaces = false

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"tests/*" = ["B008", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
